More Refactor
