+++ src/routes/Events/Detail/EventDetail.module.css
