+++ src/routes/Projects/Showcase/Showcase.jsx
@@ -1,8 +1,8 @@
-import Transition from "../../components/Transition/Transition";
-import useFetch from "../../hooks/useFetch";
-import { API_DOMAIN } from "../../utils/constants";
-import OverviewPage from "../../components/OverviewPage/OverviewPage";
-import PageTemplate from "../../components/PageTemplate/PageTemplate";
+import Transition from "../../../components/Transition/Transition";
+import useFetch from "../../../hooks/useFetch";
+import { API_DOMAIN } from "../../../utils/constants";
+import OverviewPage from "../../../components/OverviewPage/OverviewPage";
+import PageTemplate from "../../../components/PageTemplate/PageTemplate";
 
 const Projects = () => {
   const { data, isLoading } = useFetch({
