+++ src/routes/Recruitment/Ambassadors/AmbassadorDetail.jsx
@@ -1,11 +1,11 @@
 import { useParams } from "react-router-dom";
-import Typography from "../../components/Typography/Typography";
-import HeroImage from "../../components/HeroImage/HeroImage";
-import Transition from "../../components/Transition/Transition";
-import PageTemplate from "../../components/PageTemplate/PageTemplate";
-import useFetch from "../../hooks/useFetch";
-import { API_DOMAIN } from "../../utils/constants";
-import PageGap from "../../components/PageGap/PageGap";
+import Typography from "../../../components/Typography/Typography";
+import HeroImage from "../../../components/HeroImage/HeroImage";
+import Transition from "../../../components/Transition/Transition";
+import PageTemplate from "../../../components/PageTemplate/PageTemplate";
+import useFetch from "../../../hooks/useFetch";
+import { API_DOMAIN } from "../../../utils/constants";
+import PageGap from "../../../components/PageGap/PageGap";
 
 function AmbassadorDetail() {
   const params = useParams();
