+++ src/routes/Recruitment/Innotrack/Innotrack.jsx
@@ -1,22 +1,22 @@
-import Grid from "../../components/Grid/Grid";
-import Card from "../../components/PhotoCard/PhotoCard";
-import Typography from "../../components/Typography/Typography";
-import Transition from "../../components/Transition/Transition";
-import useFetch from "../../hooks/useFetch";
-import { API_DOMAIN } from "../../utils/constants";
-import PageTemplate from "../../components/PageTemplate/PageTemplate";
-import PageGap from "../../components/PageGap/PageGap";
+import Grid from "../../../components/Grid/Grid";
+import Card from "../../../components/PhotoCard/PhotoCard";
+import Typography from "../../../components/Typography/Typography";
+import Transition from "../../../components/Transition/Transition";
+import useFetch from "../../../hooks/useFetch";
+import { API_DOMAIN } from "../../../utils/constants";
+import PageTemplate from "../../../components/PageTemplate/PageTemplate";
+import PageGap from "../../../components/PageGap/PageGap";
 import {
     AccordionRoot,
     AccordionContent,
     AccordionTrigger,
     AccordionItem,
-  } from "../../components/Accordion/Accordion";
+  } from "../../../components/Accordion/Accordion";
 
 import styles from "./Innotrack.module.css";
-import Button from "../../components/Button/Button";
-import LoadingSpinner from "../../components/LoadingSpinner/LoadingSpinner";
-import Image from "../../components/Image/Image";
+import Button from "../../../components/Button/Button";
+import LoadingSpinner from "../../../components/LoadingSpinner/LoadingSpinner";
+import Image from "../../../components/Image/Image";
 
 const Innotrack = () => {
   const { data, isLoading } = useFetch({
