+++ src/routes/Recruitment/TinkeringProject/TinkeringProject.jsx
@@ -1,24 +1,24 @@
-import Grid from "../../components/Grid/Grid";
-import Card from "../../components/PhotoCard/PhotoCard";
-import Typography from "../../components/Typography/Typography";
-import Transition from "../../components/Transition/Transition";
-import useFetch from "../../hooks/useFetch";
-import { API_DOMAIN } from "../../utils/constants";
-import PageTemplate from "../../components/PageTemplate/PageTemplate";
-import PageGap from "../../components/PageGap/PageGap";
+import Grid from "../../../components/Grid/Grid";
+import Card from "../../../components/PhotoCard/PhotoCard";
+import Typography from "../../../components/Typography/Typography";
+import Transition from "../../../components/Transition/Transition";
+import useFetch from "../../../hooks/useFetch";
+import { API_DOMAIN } from "../../../utils/constants";
+import PageTemplate from "../../../components/PageTemplate/PageTemplate";
+import PageGap from "../../../components/PageGap/PageGap";
 import {
     AccordionRoot,
     AccordionContent,
     AccordionTrigger,
     AccordionItem,
-  } from "../../components/Accordion/Accordion";
+  } from "../../../components/Accordion/Accordion";
 
-import styles from "./InnovatorsTrack.module.css";
-import Button from "../../components/Button/Button";
-import LoadingSpinner from "../../components/LoadingSpinner/LoadingSpinner";
-import Image from "../../components/Image/Image";
+import styles from "./TinkeringProject.module.css";
+import Button from "../../../components/Button/Button";
+import LoadingSpinner from "../../../components/LoadingSpinner/LoadingSpinner";
+import Image from "../../../components/Image/Image";
 
-const InnovatorsTrack = () => {
+const TinkeringProjects = () => {
   const { data, isLoading } = useFetch({
     url: API_DOMAIN + "?type=tinkering",
   });
@@ -30,7 +30,7 @@ const InnovatorsTrack = () => {
           <div className={styles["content-wrapper"]}>
             <div className={styles["banner"]}>
               <div className={styles["banner-space"]}>
-                <Typography variant="banner">INNOVATOR'S TRACK</Typography>
+                <Typography variant="banner">TINKERING PROJECTS</Typography>
                 <div className={styles["scroll-more"]}>
                   <Typography
                     variant="body"
@@ -44,16 +44,14 @@ const InnovatorsTrack = () => {
               <Image
                 className={styles["banner-image"]}
                 src={data[0].bannerImage}
-                alt="Innovator's Track Cover"
+                alt="Tinkering Cover"
               />
             </div>
             <section className={styles["section-wrapper"]}>
-              {/* Intro section */}
               <Typography variant="heading">Introduction</Typography>
               <Typography variant="body"> {data[0].introduction}
               </Typography>
             
-              {/* Join section */} 
               <Typography variant="heading">How Can I Join?</Typography> 
               <div className={styles["join-text"]}> 
                   <div className={styles["text-section"]}> 
@@ -70,7 +68,6 @@ const InnovatorsTrack = () => {
                   </div> 
               </div> 
                
-              {/* Garage Assigned section */} 
               <Typography variant="heading">Alternatively,</Typography> 
               <div className={styles["garage-assigned-text"]}> 
                   <div className={styles["text-section"]}> 
@@ -102,7 +99,7 @@ const InnovatorsTrack = () => {
               )}
               </div>
               <div>
-                <Button to="/assigned_projects"  
+                <Button to="/project-openings"  
                   style={{ width: '100%', textAlign: 'center', padding: '1rem' }}
                   >
                   View All
@@ -110,9 +107,8 @@ const InnovatorsTrack = () => {
               </div>
             </section>
 
-            {/* Support section */} 
             <section className={styles["section-wrapper"]}>
-              <Typography variant={"heading"}>SUPPORT</Typography>
+              <Typography variant={"heading"}>SUPPPORT</Typography>
               {data ? (
                 <Grid desktop={3} tablet={2} mobile={1}>
                   {data.map((card, index) => (
@@ -144,12 +140,11 @@ const InnovatorsTrack = () => {
               )}
             </section>
 
-            {/* Programmes section */} 
             <section className={styles["section-wrapper"]}>
                 <Typography variant={"heading"}>PROGRAMMES</Typography>
                 <div className={styles["programme-section"]}> 
                   <div className={styles["text-section"]}> 
-                    <Typography variant={"smallHeading"}>Innovator's Night</Typography>
+                    <Typography variant={"smallHeading"}>Tinkerer's Night</Typography>
                     {data && (
                     <section className={styles["tinkererNight"]}>
                         <Image
@@ -176,7 +171,7 @@ const InnovatorsTrack = () => {
                         </div>
                     </section>
                     )}
-                    <Typography variant={"smallHeading"}>Innovator's Track Sharing Session</Typography>
+                    <Typography variant={"smallHeading"}>Tinkering Projects Sharing Session</Typography>
                     <Typography variant={"body"}>
                       {data[0].tinkererProjectDescription}
                     </Typography>
@@ -202,7 +197,6 @@ const InnovatorsTrack = () => {
               </div> 
             </section>
 
-            {/* FAQ section */} 
             <section className={styles["section-wrapper"]}>
               <Typography variant={"heading"}>FREQUENTLY ASKED QUESTIONS</Typography>
               <PageGap>
@@ -218,7 +212,7 @@ const InnovatorsTrack = () => {
                           <Typography key={i} variant="body">
                             {index === 0 && i === 1 ? (
                               <a
-                                href="/assigned_projects"
+                                href="/project-openings"
                                 target="_blank"
                                 rel="noopener noreferrer"
                                 className={styles["faq-link"]}
@@ -244,4 +238,4 @@ const InnovatorsTrack = () => {
   );
 };
 
-export default InnovatorsTrack;
+export default TinkeringProjects;
